<MegaAVR> Upgrades to platform files
- Removed file and function redundancy
- Refactored reset handling code
  - eliminated the use of switch case and employed if-else ladder
  - Implication: Saved dmem and increased usage of imem
