+++ src/platform/mega_avr/common/platform/platform_reset.c
@@ -3,11 +3,8 @@
  * Copyrights (C) 2019, Cyancore Team
  *
  * File Name		: platform_reset.c
- * Description		: This file contains sources for platform reset
- *			  dummy apis. This is necessary to avoid comilation
- *			  failure when reset handling functions arent setup.
- *			  Original file can be found at
- *			  <Platform>/platform/platform_dp.c
+ * Description		: This file contains sources for platform
+ *			  reset apis
  * Primary Author	: Akash Kollipara [akashkollipara@gmail.com]
  * Organisation		: Cyancore Core-Team
  */
@@ -15,6 +12,7 @@
 #include <stdint.h>
 #include <status.h>
 #include <platform.h>
+#include <driver/watchdog.h>
 #include <terravisor/platform.h>
 
 extern uint8_t reset_syndrome;
@@ -23,15 +21,29 @@ extern uint8_t reset_syndrome;
  * platform_get_reset_syndrome - returns the cause of reset
  *
  * @brief This function returns the information related to
- * the reset sources. Here it is a weak functiona actual
- * implementation can be found in the original file
- * mentioned in the file description.
+ * the reset sources. 
  *
  * @return reset_cause: This function returs the reset cause
  */
-_WEAK reset_t plarform_get_reset_syndrome()
+reset_t platform_get_reset_syndrome()
 {
-	return (reset_t) reset_syndrome;
+	if(reset_syndrome & 1)		/* Power on Reset */
+		return power_on_reset;
+	else if(reset_syndrome & 2)	/* External Reset */
+		return external_reset;
+	else if(reset_syndrome & 4)	/* Brown out Reset */
+		return brownout_reset;
+	else if(reset_syndrome & 8)	/* WDog Reset */
+		return wdog_reset;
+	else
+		return inval_reset;
+}
+
+
+TODO(This Function needs to be replaced with brownout detection driver api)
+_WEAK void brownout_reset_handler()
+{
+	plat_panic_handler();
 }
 
 /**
@@ -44,9 +56,15 @@ _WEAK reset_t plarform_get_reset_syndrome()
  *
  * @return void
  */
-_WEAK void platform_reset_handler(reset_t rsyn)
+void platform_reset_handler(reset_t rsyn)
 {
-	if(rsyn > external_reset)
+	if((rsyn == power_on_reset) || (rsyn == external_reset))
+		return;
+	else if(rsyn ==  brownout_reset)
+		brownout_reset_handler();
+	else if(rsyn ==  wdog_reset)
+		wdog_reset_handler();
+	else
 		plat_panic_handler();
-	return;
 }
+
